refactor(sdk): migrate AgentRuntime methods to OpenCodeHttpEngine
OpenCodeClient becomes a façade: the constructor builds an engine
parameterized by transport (baseUrl/directory/password); the 22
AgentRuntime methods forward to it. OpenCode-specific methods
(listProviders, addMcpServer, oauth*, …) and WorkspaceOps stay on
the façade. The listener API (getStatus/onEvent/onStatus) forwards too,
so store-level subscriptions land on the engine where emits happen.

Behavior preserved: 583 tests green, tsc clean. getClient() callers
unchanged. Sets up Step 3 (kimi ACP), which will add a sibling engine.
