+++ packages/sdk/src/engines/opencode-http.ts
@@ -2,70 +2,969 @@
 //
 // HTTP+SSE transport for OpenCode-class runtimes. This is the code that used
 // to live in OpenCodeClient, extracted so a future second HTTP-speaking
-// runtime can reuse it. OpenCodeClient (the façade) will compose this.
+// runtime can reuse it. OpenCodeClient (the façade) composes this.
 //
-// Migration is incremental (Task 2.3): methods move here one at a time, each
-// committed with its test. Until migration is complete, this class throws on
-// every method; once complete, OpenCodeClient forwards to it.
+// The engine owns the wire: fetch/SSE transport, auth headers, request
+// timeouts, event-stream normalization, and the 22 AgentRuntime methods. The
+// façade (OpenCodeClient) constructs one and forwards to it; OpenCode-specific
+// configuration methods (provider/MCP/OAuth) on the façade borrow the engine's
+// HTTP helpers (headers/fetchWithTimeout/apiError/dirQuery/disposeInstance).
 
 import type { TransportSpec } from "../runtime-def";
 import { BaseAgentRuntime } from "../base-runtime";
-import type {
-  AgentInfo, CommandInfo, HistoryMessage, OpenCodeEvent,
-  PermissionAskedEvent, PermissionReply, QuestionAskedEvent,
-  RuntimeStatus, SessionMeta, SkillInfo,
+import type { AgentRuntime } from "../runtime";
+import {
+  type AgentInfo,
+  type CommandInfo,
+  type HistoryMessage,
+  type OpenCodePart,
+  type OpenCodeRawEvent,
+  type PermissionAskedEvent,
+  type PermissionReply,
+  type QuestionAskedEvent,
+  type SessionMeta,
+  type SkillInfo,
+  type ToolCallStatus,
 } from "../types";
 
 export type OpenCodeHttpTransport = Extract<TransportSpec, { kind: "opencode-http" }>;
 
+/** Engine options beyond what the RuntimeDef transport carries. These are the
+ *  runtime-level knobs the host (OpenCodeClient) owns: injected fetch for
+ *  tests, the SSE-handshake / short-request timeouts, and the basic-auth
+ *  username (password rides on the transport). The engine itself reads ONLY
+ *  baseUrl/directory/password from `transport`; everything else comes from
+ *  here, so a second HTTP runtime can pass its own. */
+export interface OpenCodeHttpEngineOptions {
+  fetchImpl?: typeof fetch;
+  connectTimeoutMs?: number;
+  requestTimeoutMs?: number;
+  username?: string;
+}
+
+function mapToolStatus(status: string): ToolCallStatus {
+  switch (status) {
+    case "running":
+      return "running";
+    case "completed":
+      return "success";
+    case "error":
+      return "failed";
+    default:
+      return "pending";
+  }
+}
+
+/** OpenCode's error objects nest the human-readable message at
+ *  error.data.message; keep the first line (some errors append a stack). */
+function errorText(error: unknown): string | undefined {
+  const err = error as { name?: string; message?: string; data?: { message?: string } } | undefined;
+  const full = err?.data?.message ?? err?.message ?? err?.name;
+  return typeof full === "string" && full ? full.split("\n")[0] : undefined;
+}
+
+/** Split a "provider/model" default-model string into the `{providerID, modelID}`
+ *  shape the prompt API expects. Returns undefined when the string is empty or
+ *  has no provider prefix (so the caller omits the key and the server falls back
+ *  to its own default). Splits on the FIRST slash: a model id may contain more. */
+function parseModel(model?: string | null): { providerID: string; modelID: string } | undefined {
+  if (!model) return undefined;
+  const i = model.indexOf("/");
+  if (i <= 0 || i >= model.length - 1) return undefined;
+  return { providerID: model.slice(0, i), modelID: model.slice(i + 1) };
+}
+
 export class OpenCodeHttpEngine extends BaseAgentRuntime {
-  constructor(private transport: OpenCodeHttpTransport) {
+  private readonly _baseUrl: string;
+  private readonly fetchImpl: typeof fetch;
+  private readonly authHeader: string | null;
+  /** Base64 `user:password` for `?auth_token=` — the EventSource cannot set
+   *  headers, and the server accepts the same Basic payload as a query param. */
+  private readonly authToken: string | null;
+  private readonly directory: string | null;
+  private readonly customFetch: boolean;
+  private readonly connectTimeoutMs: number;
+  private readonly requestTimeoutMs: number;
+  private abort: AbortController | null = null;
+  private es: EventSource | null = null;
+  /** Pending self-heal of a dead event stream (see reconnectSoon). */
+  private reconnectTimer: ReturnType<typeof setTimeout> | null = null;
+  /** True between close() and the next connect() — stops self-healing. */
+  private closed = false;
+  /** messageID → role, learned from message.updated, to skip echoed user parts. */
+  private readonly roles = new Map<string, string>();
+  /** partID → accumulated text of a streaming text part. OpenCode publishes the
+   *  full part only at text-start (empty) and text-end; every token in between
+   *  arrives as a message.part.delta that must be summed here — otherwise the
+   *  app shows nothing until the whole passage is finished. */
+  private readonly textStreams = new Map<string, { sessionId: string; text: string }>();
+  /** partID → accumulated reasoning text. Reasoning parts stream the same way as
+   *  text (field "text" deltas) but were dropped because they were never seeded
+   *  here — so the model's thinking never reached the UI. */
+  private readonly reasoningStreams = new Map<string, { sessionId: string; text: string }>();
+  /** sessionId → the step-start part ids seen this turn. Its size is the current
+   *  step number; cleared on session.idle. Deduped because a part can update more
+   *  than once. */
+  private readonly stepParts = new Map<string, Set<string>>();
+
+  constructor(
+    transport: OpenCodeHttpTransport,
+    opts: OpenCodeHttpEngineOptions = {},
+  ) {
     super();
+    this._baseUrl = transport.baseUrl.replace(/\/$/, "");
+    this.customFetch = !!opts.fetchImpl;
+    // Bind to globalThis — an unbound `fetch` reference throws "Illegal invocation" in browsers.
+    this.fetchImpl = (opts.fetchImpl ?? globalThis.fetch).bind(globalThis);
+    const username = opts.username ?? "opencode";
+    this.authToken = transport.password ? btoa(`${username}:${transport.password}`) : null;
+    this.authHeader = this.authToken ? `Basic ${this.authToken}` : null;
+    this.directory = transport.directory ?? null;
+    this.connectTimeoutMs = opts.connectTimeoutMs ?? 5000;
+    this.requestTimeoutMs = opts.requestTimeoutMs ?? 15000;
   }
 
-  // Placeholders — replaced in Task 2.3, one method per step. Each throws so
-  // accidental use during migration fails loudly rather than silently no-op'ing.
+  // ---- HTTP helpers (shared with the OpenCodeClient façade) ----
+  // These stay public so the façade's OpenCode-specific methods (provider/MCP/
+  // OAuth config, which are deliberately OUTSIDE the AgentRuntime interface)
+  // can talk to the server through the SAME transport — one auth header, one
+  // timeout policy, one fetch impl. The façade owns NO transport state of its own.
+
+  /** The transport's base URL (trailing slash stripped). Exposed so the
+   *  façade's OpenCode-specific methods can build endpoint paths. */
+  baseUrl(): string {
+    return this._baseUrl;
+  }
+
+  /** Build the standard request headers (Content-Type + Authorization). */
+  headers(json = false): Record<string, string> {
+    const h: Record<string, string> = {};
+    if (json) h["Content-Type"] = "application/json";
+    if (this.authHeader) h["Authorization"] = this.authHeader;
+    return h;
+  }
+
+  /** fetch with an abort-driven timeout. Used by the AgentRuntime methods that
+   *  need bounded waits (createSession, getMessages, sendPrompt, listQuestions,
+   *  listPermissions, connect's fetch fallback). */
+  async fetchWithTimeout(
+    input: RequestInfo | URL,
+    init: RequestInit,
+    timeoutMs = this.requestTimeoutMs,
+  ): Promise<Response> {
+    const controller = new AbortController();
+    const timer = setTimeout(() => controller.abort(), timeoutMs);
+    try {
+      return await this.fetchImpl(input, { ...init, signal: controller.signal });
+    } catch (err) {
+      if (controller.signal.aborted) throw new Error("Timed out waiting for OpenCode");
+      throw err;
+    } finally {
+      clearTimeout(timer);
+    }
+  }
+
+  /** Issue a plain fetch (no per-call timeout) for endpoints that need custom
+   *  abort behavior (e.g. OAuth callbacks waiting on a browser redirect). */
+  fetch(input: RequestInfo | URL, init: RequestInit): Promise<Response> {
+    return this.fetchImpl(input, init);
+  }
+
+  /** Build an Error carrying the server's diagnostic message, when it has one
+   *  (OpenCode errors look like `{ name, data: { message } }`). */
+  async apiError(res: Response, what: string): Promise<Error> {
+    let detail = "";
+    try {
+      const body = (await res.json()) as { data?: { message?: string }; message?: string };
+      detail = body.data?.message ?? body.message ?? "";
+    } catch {
+      /* not JSON — keep the status alone */
+    }
+    return new Error(`${what} (${res.status}${detail ? `: ${detail}` : ""})`);
+  }
+
+  /** Append `?directory=<path>` so the server resolves the workspace instance.
+   *  Note: the `workspace` param is a `wrk_` id, NOT a path — omit it (directory
+   *  alone resolves the instance; sending a path as workspace 500s the server). */
+  dirQuery(): string {
+    return this.directory ? `?directory=${encodeURIComponent(this.directory)}` : "";
+  }
+
+  /** The server caches its provider list per instance — a credential change
+   *  (PUT/DELETE /auth, OAuth) is invisible to /config/providers until the
+   *  instance is disposed. Two instances can hold the stale cache: the default
+   *  one (answering the unscoped /config/providers this client reads) and the
+   *  workspace one (running this folder's sessions) — dispose both.
+   *  Best-effort: the credential is already stored. */
+  async disposeInstance(): Promise<void> {
+    for (const q of new Set(["", this.dirQuery()])) {
+      await this.fetchImpl(`${this._baseUrl}/instance/dispose${q}`, {
+        method: "POST",
+        headers: this.headers(true),
+        body: "{}",
+      }).catch(() => undefined);
+    }
+  }
 
   // ---- lifecycle ----
-  // getStatus / onStatus / onEvent are inherited from BaseAgentRuntime and
-  // already correct (they are protocol-agnostic plumbing). Only connect and
-  // close are HTTP-specific and need migration.
-  async connect(): Promise<void> { throw new Error("OpenCodeHttpEngine.connect: not migrated yet"); }
-  close(): void { throw new Error("OpenCodeHttpEngine.close: not migrated yet"); }
+
+  /** Open the SSE event stream. Resolves once the server acknowledges. */
+  async connect(): Promise<void> {
+    this.closed = false;
+    if (this.reconnectTimer) {
+      clearTimeout(this.reconnectTimer);
+      this.reconnectTimer = null;
+    }
+    this.setStatus("connecting");
+
+    // Prefer EventSource in a real webview/browser (reliable SSE, incl. macOS
+    // WKWebView) — auth rides along as ?auth_token=, since EventSource cannot
+    // set headers. Fall back to streaming fetch for node/tests.
+    const canUseEventSource = !this.customFetch && typeof EventSource !== "undefined";
+    if (canUseEventSource) {
+      return new Promise((resolve, reject) => {
+        let opened = false;
+        let finished = false;
+        const es = new EventSource(this.eventUrl());
+        this.es = es;
+        const timer = setTimeout(() => {
+          if (opened || finished) return;
+          finished = true;
+          this.setStatus("error");
+          es.close();
+          if (this.es === es) this.es = null;
+          reject(new Error("Timed out opening OpenCode event stream"));
+        }, this.connectTimeoutMs);
+        es.onopen = () => {
+          if (finished) return;
+          opened = true;
+          finished = true;
+          clearTimeout(timer);
+          this.setStatus("ready");
+          resolve();
+        };
+        es.onmessage = (ev) => {
+          try {
+            this.normalize(JSON.parse(ev.data) as OpenCodeRawEvent);
+          } catch {
+            /* ignore malformed frame */
+          }
+        };
+        es.onerror = () => {
+          if (!opened) {
+            if (finished) return;
+            finished = true;
+            clearTimeout(timer);
+            this.setStatus("error");
+            es.close();
+            this.es = null;
+            reject(new Error("Could not open OpenCode event stream"));
+          } else {
+            // Take over reconnection ourselves. EventSource's built-in retry
+            // never recovers when the server ends the stream terminally —
+            // e.g. the global-config PATCH (model switch) rebuilds the
+            // instance and closes /event moments AFTER acknowledging, killing
+            // even a freshly reopened stream; the app then sat in
+            // "connecting" until a manual Connect. Reopen a fresh stream
+            // with backoff instead (a deliberate close() cancels it).
+            es.close();
+            if (this.es === es) {
+              this.es = null;
+              this.reconnectSoon();
+            }
+          }
+        };
+      });
+    }
+
+    this.abort = new AbortController();
+    return new Promise((resolve, reject) => {
+      let opened = false;
+      const abort = this.abort!;
+      const timer = setTimeout(() => {
+        if (!opened) abort.abort(new Error("Timed out opening OpenCode event stream"));
+      }, this.connectTimeoutMs);
+      this.fetchImpl(this.eventUrl(), {
+        headers: { Accept: "text/event-stream", ...this.headers() },
+        signal: abort.signal,
+      })
+        .then(async (res) => {
+          clearTimeout(timer);
+          if (!res.ok || !res.body) {
+            this.setStatus("error");
+            reject(new Error(`OpenCode /event returned ${res.status}`));
+            return;
+          }
+          this.setStatus("ready");
+          opened = true;
+          resolve();
+          await this.readStream(res.body);
+        })
+        .catch((err) => {
+          clearTimeout(timer);
+          if (!opened) {
+            this.setStatus("error");
+            reject(err instanceof Error ? err : new Error(String(err)));
+          } else {
+            this.setStatus("offline");
+          }
+        });
+    });
+  }
+
+  close(): void {
+    this.closed = true;
+    if (this.reconnectTimer) {
+      clearTimeout(this.reconnectTimer);
+      this.reconnectTimer = null;
+    }
+    this.es?.close();
+    this.es = null;
+    this.abort?.abort();
+    this.abort = null;
+    this.setStatus("offline");
+  }
+
+  /** Reopen the event stream after it died post-open, with backoff. The first
+   *  retry is quick (a plain blip); later ones stretch to cover the server's
+   *  instance-rebuild window after a global-config PATCH. Gives up into
+   *  "error" so the banner offers the manual Connect. */
+  private reconnectSoon(attempt = 0): void {
+    if (this.closed || this.reconnectTimer) return;
+    this.setStatus("connecting");
+    const delay = attempt === 0 ? 250 : Math.min(1000 * attempt, 3000);
+    this.reconnectTimer = setTimeout(() => {
+      this.reconnectTimer = null;
+      if (this.closed) return;
+      this.connect().catch(() => {
+        if (attempt + 1 < 8) this.reconnectSoon(attempt + 1);
+        else this.setStatus("error");
+      });
+    }, delay);
+  }
 
   // ---- sessions ----
-  async createSession(): Promise<string> { throw new Error("OpenCodeHttpEngine.createSession: not migrated yet"); }
-  async listSessions(): Promise<SessionMeta[]> { throw new Error("OpenCodeHttpEngine.listSessions: not migrated yet"); }
-  async deleteSession(_sessionId: string): Promise<void> { throw new Error("OpenCodeHttpEngine.deleteSession: not migrated yet"); }
-  async getMessages(_sessionId: string): Promise<HistoryMessage[]> { throw new Error("OpenCodeHttpEngine.getMessages: not migrated yet"); }
-  async sendPrompt(_sessionId: string, _text: string, _agent?: string, _model?: string | null): Promise<void> { throw new Error("OpenCodeHttpEngine.sendPrompt: not migrated yet"); }
-  async abortSession(_sessionId: string): Promise<void> { throw new Error("OpenCodeHttpEngine.abortSession: not migrated yet"); }
-  async revert(_sessionId: string, _messageID: string, _partID?: string): Promise<void> { throw new Error("OpenCodeHttpEngine.revert: not migrated yet"); }
-  async unrevert(_sessionId: string): Promise<void> { throw new Error("OpenCodeHttpEngine.unrevert: not migrated yet"); }
+
+  /** Create a new agent session, returning its id. Scoping is by the sidecar's
+   *  working directory (set at spawn), not a query param — passing `?directory=`
+   *  here routes the turn to a scope whose events the global stream never sees. */
+  async createSession(): Promise<string> {
+    // The directory decides where the session lives and works — without it the
+    // server would put it in the process's boot folder, not the active one.
+    const res = await this.fetchWithTimeout(`${this._baseUrl}/session${this.dirQuery()}`, {
+      method: "POST",
+      headers: this.headers(true),
+      body: "{}",
+    });
+    if (!res.ok) throw await this.apiError(res, "Failed to create session");
+    const json = (await res.json()) as { id: string };
+    return json.id;
+  }
+
+  /** List existing sessions (conversation history), newest first — across ALL
+   *  workspace folders. The plain `/session` list is scoped to the project the
+   *  sidecar's cwd resolves to, so history would appear to change when the user
+   *  switches folders; `/experimental/session` lists every project's sessions
+   *  (each item still carries its `directory`). The OpenCode version is pinned,
+   *  so the experimental route is stable for us; fall back to `/session` if a
+   *  server ever lacks it. */
+  async listSessions(): Promise<SessionMeta[]> {
+    let res = await this.fetchWithTimeout(`${this._baseUrl}/experimental/session`, {
+      headers: this.headers(),
+    });
+    if (!res.ok) {
+      res = await this.fetchWithTimeout(`${this._baseUrl}/session`, { headers: this.headers() });
+    }
+    if (!res.ok) throw await this.apiError(res, "Failed to list sessions");
+    const arr = (await res.json()) as Array<{
+      id: string;
+      title?: string;
+      slug?: string;
+      directory?: string;
+      parentID?: string | null;
+      time?: { created?: number; updated?: number };
+    }>;
+    return arr.map((s) => ({
+      id: s.id,
+      title: s.title ?? "Untitled",
+      slug: s.slug,
+      directory: s.directory,
+      parentId: s.parentID ?? undefined,
+      created: s.time?.created,
+      updated: s.time?.updated,
+    }));
+  }
+
+  /** Delete a session. */
+  async deleteSession(sessionId: string): Promise<void> {
+    const res = await this.fetchImpl(`${this._baseUrl}/session/${encodeURIComponent(sessionId)}`, {
+      method: "DELETE",
+      headers: this.headers(),
+    });
+    if (!res.ok) throw await this.apiError(res, "Failed to delete session");
+  }
+
+  /** Load a session's message history. */
+  async getMessages(sessionId: string): Promise<HistoryMessage[]> {
+    const res = await this.fetchWithTimeout(
+      `${this._baseUrl}/session/${encodeURIComponent(sessionId)}/message`,
+      { headers: this.headers() },
+    );
+    if (!res.ok) throw await this.apiError(res, "Failed to load messages");
+    const arr = (await res.json()) as Array<{
+      info: {
+        id?: string;
+        role: "user" | "assistant";
+        time?: { completed?: number };
+        error?: unknown;
+        agent?: string;
+      };
+      parts: HistoryMessage["parts"];
+    }>;
+    return arr.map((m) => {
+      const error = errorText(m.info.error);
+      return {
+        role: m.info.role,
+        ...(m.info.id ? { id: m.info.id } : {}),
+        completed: m.info.time?.completed,
+        ...(error ? { error } : {}),
+        ...(m.info.agent ? { agent: m.info.agent } : {}),
+        parts: m.parts ?? [],
+      };
+    });
+  }
+
+  /** Revert the session to (and including) a message: OpenCode drops that
+   *  message and everything after it, rolling back any workspace files it
+   *  touched. This is how the app edits a past user message — revert to it,
+   *  then send the corrected text. The session must be idle (abort first);
+   *  reverting a busy session is rejected by OpenCode. `partID` reverts to a
+   *  specific part within the message; omit to revert the whole message. */
+  async revert(sessionId: string, messageID: string, partID?: string): Promise<void> {
+    const res = await this.fetchImpl(
+      `${this._baseUrl}/session/${encodeURIComponent(sessionId)}/revert`,
+      {
+        method: "POST",
+        headers: this.headers(true),
+        body: JSON.stringify({ messageID, ...(partID ? { partID } : {}) }),
+      },
+    );
+    if (!res.ok) throw await this.apiError(res, "Failed to revert the message");
+  }
+
+  /** Undo the last revert — restores the reverted messages and files. */
+  async unrevert(sessionId: string): Promise<void> {
+    const res = await this.fetchImpl(
+      `${this._baseUrl}/session/${encodeURIComponent(sessionId)}/unrevert`,
+      { method: "POST", headers: this.headers(true), body: "{}" },
+    );
+    if (!res.ok) throw await this.apiError(res, "Failed to restore the reverted messages");
+  }
+
+  /** Interrupt the session's current turn (POST /session/:id/abort). A no-op
+   *  on an idle session — the server just answers false. */
+  async abortSession(sessionId: string): Promise<void> {
+    const res = await this.fetchImpl(
+      `${this._baseUrl}/session/${encodeURIComponent(sessionId)}/abort`,
+      { method: "POST", headers: this.headers(true), body: "{}" },
+    );
+    if (!res.ok) throw await this.apiError(res, "Failed to interrupt the session");
+  }
 
   // ---- capability discovery ----
-  async listSkills(): Promise<SkillInfo[]> { throw new Error("OpenCodeHttpEngine.listSkills: not migrated yet"); }
-  async listAgents(): Promise<AgentInfo[]> { throw new Error("OpenCodeHttpEngine.listAgents: not migrated yet"); }
-  async listCommands(): Promise<CommandInfo[]> { throw new Error("OpenCodeHttpEngine.listCommands: not migrated yet"); }
+
+  /** Real skills loaded by OpenCode (built-in + bundled + user). */
+  async listSkills(): Promise<SkillInfo[]> {
+    // Scope to the workspace: skill instances are created lazily per directory,
+    // and the unscoped endpoint answers from an instance that may have none.
+    const query = this.directory ? `?directory=${encodeURIComponent(this.directory)}` : "";
+    const res = await this.fetchImpl(`${this._baseUrl}/api/skill${query}`, {
+      headers: this.headers(),
+    });
+    if (!res.ok) throw await this.apiError(res, "Failed to list skills");
+    const body = (await res.json()) as { data?: SkillInfo[] };
+    return body.data ?? [];
+  }
+
+  /** Real agents configured in OpenCode. */
+  async listAgents(): Promise<AgentInfo[]> {
+    const res = await this.fetchImpl(`${this._baseUrl}/agent`, { headers: this.headers() });
+    if (!res.ok) throw await this.apiError(res, "Failed to list agents");
+    return (await res.json()) as AgentInfo[];
+  }
+
+  /** Slash commands the runtime can run — config commands, skills and MCP
+   *  prompts all surface in this one list (directory-scoped like skills). */
+  async listCommands(): Promise<CommandInfo[]> {
+    const res = await this.fetchImpl(`${this._baseUrl}/command${this.dirQuery()}`, {
+      headers: this.headers(),
+    });
+    if (!res.ok) throw await this.apiError(res, "Failed to list commands");
+    const arr = (await res.json()) as Array<{
+      name: string;
+      description?: string;
+      source?: string;
+      agent?: string;
+      // A string for config commands and skills; MCP prompts report an
+      // argument-schema OBJECT here — only a string is a usable template.
+      template?: unknown;
+    }>;
+    return arr.map((c) => ({
+      name: c.name,
+      description: c.description,
+      source: c.source,
+      agent: c.agent,
+      template: typeof c.template === "string" ? c.template : undefined,
+    }));
+  }
 
   // ---- model selection ----
-  async getDefaultModel(): Promise<string | null> { throw new Error("OpenCodeHttpEngine.getDefaultModel: not migrated yet"); }
-  async setDefaultModel(_model: string): Promise<void> { throw new Error("OpenCodeHttpEngine.setDefaultModel: not migrated yet"); }
+
+  /** The configured default model ("provider/model"), or null when unset. */
+  async getDefaultModel(): Promise<string | null> {
+    // Read the same global config that setDefaultModel PATCHes. The instance-
+    // scoped /config only reflects a model change after OpenCode rebuilds the
+    // instance (~1s later), so reading it right after a switch returns the
+    // previous model.
+    const res = await this.fetchImpl(`${this._baseUrl}/global/config`, { headers: this.headers() });
+    if (!res.ok) throw await this.apiError(res, "Failed to read config");
+    const cfg = (await res.json()) as { model?: string };
+    return cfg.model ?? null;
+  }
+
+  /** Set the default model in OpenCode's global (app-profile) config. */
+  async setDefaultModel(model: string): Promise<void> {
+    const res = await this.fetchImpl(`${this._baseUrl}/global/config`, {
+      method: "PATCH",
+      headers: this.headers(true),
+      body: JSON.stringify({ model }),
+    });
+    if (!res.ok) throw await this.apiError(res, "Failed to set model");
+    // NOTE: deliberately do NOT disposeInstance() here (unlike auth changes).
+    // Disposing tears the instance down asynchronously server-side, landing
+    // ~1s AFTER the store's connectRetry has already opened a fresh event
+    // stream — it kills that stream and EventSource cannot recover, stranding
+    // the app in "connecting" until a manual Connect. The model switch is
+    // applied by the config PATCH; the reconnect picks it up.
+  }
 
   // ---- agent-driven execution ----
-  async runShell(_sessionId: string, _command: string, _agent?: string): Promise<void> { throw new Error("OpenCodeHttpEngine.runShell: not migrated yet"); }
-  async runCommand(_sessionId: string, _command: string, _args?: string): Promise<void> { throw new Error("OpenCodeHttpEngine.runCommand: not migrated yet"); }
 
-  // ---- interactive requests ----
-  async listQuestions(_sessionId?: string): Promise<QuestionAskedEvent[]> { throw new Error("OpenCodeHttpEngine.listQuestions: not migrated yet"); }
-  async listPermissions(_sessionId?: string): Promise<PermissionAskedEvent[]> { throw new Error("OpenCodeHttpEngine.listPermissions: not migrated yet"); }
-  async answerQuestion(_requestId: string, _answers: string[][]): Promise<void> { throw new Error("OpenCodeHttpEngine.answerQuestion: not migrated yet"); }
-  async rejectQuestion(_requestId: string): Promise<void> { throw new Error("OpenCodeHttpEngine.rejectQuestion: not migrated yet"); }
-  async replyPermission(_requestId: string, _reply: PermissionReply): Promise<void> { throw new Error("OpenCodeHttpEngine.replyPermission: not migrated yet"); }
+  /** Run a shell command directly in the session's workspace — no model turn.
+   *  The result lands in the session history as a bash tool part and streams
+   *  via onEvent; the POST resolves only when the command finishes. */
+  async runShell(sessionId: string, command: string, agent = "build"): Promise<void> {
+    const res = await this.fetchImpl(
+      `${this._baseUrl}/session/${encodeURIComponent(sessionId)}/shell`,
+      {
+        method: "POST",
+        headers: this.headers(true),
+        body: JSON.stringify({ agent, command }),
+      },
+    );
+    if (!res.ok) throw await this.apiError(res, "Command failed to run");
+  }
+
+  /** Run a slash command (config command / skill / MCP prompt) in a session.
+   *  This is a full agent turn; the POST resolves when the turn completes,
+   *  while output streams via onEvent along the way. */
+  async runCommand(sessionId: string, command: string, args?: string): Promise<void> {
+    const res = await this.fetchImpl(
+      `${this._baseUrl}/session/${encodeURIComponent(sessionId)}/command`,
+      {
+        method: "POST",
+        headers: this.headers(true),
+        body: JSON.stringify({ command, ...(args ? { arguments: args } : {}) }),
+      },
+    );
+    if (!res.ok) throw await this.apiError(res, `Failed to run /${command}`);
+  }
+
+  /** Send a prompt into a session; output streams back via onEvent (SSE).
+   *  `agent` pins the turn to a specific agent (e.g. "plan"); omitted, the
+   *  server resolves its default — the key is left out entirely so build-mode
+   *  requests stay byte-identical to before.
+   *
+   *  `model` ("provider/model") pins the turn to the current default model.
+   *  OpenCode persists a `session.model` at session creation, so switching the
+   *  global default does NOT rebind existing sessions — an old session would
+   *  keep answering on its creation-time provider/model. Passing the current
+   *  default on every turn overrides that stale binding without mutating the
+   *  server's stored rows. Omitted/unparseable, the key is left out and the
+   *  server falls back to the session/global default. */
+  async sendPrompt(
+    sessionId: string,
+    text: string,
+    agent?: string,
+    model?: string | null,
+  ): Promise<void> {
+    const m = parseModel(model);
+    const res = await this.fetchWithTimeout(
+      `${this._baseUrl}/session/${encodeURIComponent(sessionId)}/prompt_async`,
+      {
+        method: "POST",
+        headers: this.headers(true),
+        body: JSON.stringify({
+          parts: [{ type: "text", text }],
+          ...(agent ? { agent } : {}),
+          ...(m ? { model: m } : {}),
+        }),
+      },
+    );
+    if (!res.ok) throw await this.apiError(res, "Failed to send prompt");
+  }
+
+  // ---- interactive requests (question / permission) ----
+  // OpenCode exposes these as directory-scoped GLOBAL lists (not session-nested):
+  // GET/POST /question[/…] and /permission[/…], each scoped by ?directory=. The
+  // request id is globally unique; the reply/reject endpoints take it directly.
+
+  /** Pending questions in the workspace (recovery on open — an ask can predate connect). */
+  async listQuestions(_sessionId?: string): Promise<QuestionAskedEvent[]> {
+    const res = await this.fetchWithTimeout(`${this._baseUrl}/question${this.dirQuery()}`, {
+      headers: this.headers(),
+    });
+    if (!res.ok) return [];
+    const arr = (await res.json()) as Array<{
+      id: string;
+      sessionID: string;
+      questions?: QuestionAskedEvent["questions"];
+    }>;
+    return arr.map((q) => ({
+      type: "question.asked" as const,
+      sessionId: q.sessionID,
+      requestId: q.id,
+      questions: q.questions ?? [],
+    }));
+  }
+
+  /** Answer a question: one array of selected option labels per question, in order. */
+  async answerQuestion(requestId: string, answers: string[][]): Promise<void> {
+    const res = await this.fetchImpl(
+      `${this._baseUrl}/question/${encodeURIComponent(requestId)}/reply${this.dirQuery()}`,
+      { method: "POST", headers: this.headers(true), body: JSON.stringify({ answers }) },
+    );
+    if (!res.ok) throw await this.apiError(res, "Failed to answer the question");
+  }
+
+  /** Reject/dismiss a question (the agent proceeds without an answer). */
+  async rejectQuestion(requestId: string): Promise<void> {
+    const res = await this.fetchImpl(
+      `${this._baseUrl}/question/${encodeURIComponent(requestId)}/reject${this.dirQuery()}`,
+      { method: "POST", headers: this.headers(true), body: "{}" },
+    );
+    if (!res.ok) throw await this.apiError(res, "Failed to reject the question");
+  }
+
+  /** Pending permission requests in the workspace (recovery on open). */
+  async listPermissions(_sessionId?: string): Promise<PermissionAskedEvent[]> {
+    const res = await this.fetchWithTimeout(`${this._baseUrl}/permission${this.dirQuery()}`, {
+      headers: this.headers(),
+    });
+    if (!res.ok) return [];
+    // Same dual field names as the SSE event: `permission`/`patterns` (V2)
+    // with `action`/`resources` as the legacy fallback.
+    const arr = (await res.json()) as Array<{
+      id: string;
+      sessionID: string;
+      permission?: string;
+      patterns?: string[];
+      action?: string;
+      resources?: string[];
+    }>;
+    return arr.map((p) => ({
+      type: "permission.asked" as const,
+      sessionId: p.sessionID,
+      requestId: p.id,
+      action: p.permission ?? p.action ?? "action",
+      resources: p.patterns ?? p.resources ?? [],
+    }));
+  }
+
+  /** Reply to a permission request: allow once, allow always, or reject. */
+  async replyPermission(requestId: string, reply: PermissionReply): Promise<void> {
+    const res = await this.fetchImpl(
+      `${this._baseUrl}/permission/${encodeURIComponent(requestId)}/reply${this.dirQuery()}`,
+      { method: "POST", headers: this.headers(true), body: JSON.stringify({ reply }) },
+    );
+    if (!res.ok) throw await this.apiError(res, "Failed to reply to the permission");
+  }
+
+  // ---- internals ----
+
+  /** The /event stream URL: directory scope + auth_token (EventSource has no headers). */
+  private eventUrl(): string {
+    const params = new URLSearchParams();
+    if (this.directory) params.set("directory", this.directory);
+    if (this.authToken) params.set("auth_token", this.authToken);
+    const q = params.toString();
+    return `${this._baseUrl}/event${q ? `?${q}` : ""}`;
+  }
+
+  private async readStream(body: ReadableStream<Uint8Array>): Promise<void> {
+    const reader = body.getReader();
+    const decoder = new TextDecoder();
+    let buffer = "";
+    try {
+      for (;;) {
+        const { done, value } = await reader.read();
+        if (done) break;
+        buffer += decoder.decode(value, { stream: true });
+        let sep: number;
+        while ((sep = buffer.indexOf("\n\n")) !== -1) {
+          const chunk = buffer.slice(0, sep);
+          buffer = buffer.slice(sep + 2);
+          this.handleSseChunk(chunk);
+        }
+      }
+    } catch {
+      // aborted or connection dropped
+    } finally {
+      this.setStatus("offline");
+    }
+  }
+
+  private handleSseChunk(chunk: string): void {
+    const dataLines = chunk
+      .split("\n")
+      .filter((l) => l.startsWith("data:"))
+      .map((l) => l.slice(5).trim());
+    if (dataLines.length === 0) return;
+    let raw: OpenCodeRawEvent;
+    try {
+      raw = JSON.parse(dataLines.join("\n"));
+    } catch {
+      return;
+    }
+    this.normalize(raw);
+  }
+
+  private normalize(raw: OpenCodeRawEvent): void {
+    const props = raw.properties ?? {};
+    switch (raw.type) {
+      case "message.updated": {
+        // Learn each message's role so we can skip the echoed user message parts.
+        const info = props.info as
+          | { id?: string; role?: string; sessionID?: string; agent?: string }
+          | undefined;
+        if (info?.id && info.role) this.roles.set(info.id, info.role);
+        // A user message surfaces its id (so the app can tag the live block for
+        // editing) and its agent (so the per-session agent mode stays in sync —
+        // plan_exit "Yes" injects a build one). Emitted for every user message,
+        // carrying whichever fields are present.
+        if (info?.role === "user" && info.sessionID && (info.id || typeof info.agent === "string")) {
+          this.emit({
+            type: "message.agent",
+            sessionId: String(info.sessionID),
+            ...(info.id ? { messageID: info.id } : {}),
+            ...(typeof info.agent === "string" ? { agent: info.agent } : {}),
+          });
+        }
+        break;
+      }
+      case "message.part.updated": {
+        const part = props.part as
+          | (OpenCodePart & { sessionID?: string; messageID?: string })
+          | undefined;
+        if (!part) return;
+        // The user's own message is echoed here; the app already shows it locally.
+        if (part.messageID && this.roles.get(String(part.messageID)) === "user") return;
+        const sessionId = String(part.sessionID ?? "");
+        if (part.type === "text") {
+          const t = part as { id: string; text: string };
+          this.textStreams.set(t.id, { sessionId, text: t.text ?? "" });
+          this.emit({ type: "text.updated", sessionId, partId: t.id, text: t.text ?? "" });
+        } else if (part.type === "reasoning") {
+          // Seed the reasoning stream so its "text" deltas accumulate (below),
+          // and surface the thinking the app used to discard.
+          const r = part as unknown as { id: string; text?: string };
+          this.reasoningStreams.set(r.id, { sessionId, text: r.text ?? "" });
+          this.emit({ type: "reasoning.updated", sessionId, partId: r.id, text: r.text ?? "" });
+        } else if (part.type === "step-start") {
+          // A new model step began — count it (deduped) so the UI can show
+          // "step N" and prove the turn is progressing rather than hung.
+          const sp = part as unknown as { id: string };
+          let seen = this.stepParts.get(sessionId);
+          if (!seen) this.stepParts.set(sessionId, (seen = new Set()));
+          if (!seen.has(sp.id)) {
+            seen.add(sp.id);
+            this.emit({ type: "step.updated", sessionId, step: seen.size });
+          }
+        } else if (part.type === "tool") {
+          const tp = part as {
+            callID: string;
+            tool: string;
+            state?: {
+              status?: string;
+              title?: string;
+              input?: Record<string, unknown>;
+              output?: string;
+              time?: { start?: number; end?: number };
+              metadata?: { sessionId?: unknown; output?: unknown; diff?: unknown };
+            };
+          };
+          // A task tool's metadata names the subagent session it spawned.
+          const meta = tp.state?.metadata;
+          const child = meta?.sessionId;
+          this.emit({
+            type: "tool.updated",
+            sessionId,
+            callId: tp.callID,
+            tool: tp.tool,
+            status: mapToolStatus(tp.state?.status ?? "pending"),
+            title: tp.state?.title,
+            input: tp.state?.input,
+            output: typeof tp.state?.output === "string" ? tp.state.output : undefined,
+            // While running, bash accumulates its stdout tail here — the live
+            // output the UI shows so a long command never looks hung.
+            partialOutput: typeof meta?.output === "string" ? meta.output : undefined,
+            diff: typeof meta?.diff === "string" ? meta.diff : undefined,
+            startedAt: typeof tp.state?.time?.start === "number" ? tp.state.time.start : undefined,
+            endedAt: typeof tp.state?.time?.end === "number" ? tp.state.time.end : undefined,
+            childSessionId: typeof child === "string" ? child : undefined,
+          });
+        }
+        break;
+      }
+      case "message.part.delta": {
+        // One streamed token. Both text and reasoning parts stream via the
+        // "text" field; each was seeded above, so route the delta to whichever
+        // stream owns this part.
+        const d = props as { partID?: string; field?: string; delta?: string };
+        if (d.field !== "text" || !d.partID || typeof d.delta !== "string") return;
+        const partId = String(d.partID);
+        const acc = this.textStreams.get(partId);
+        if (acc) {
+          acc.text += d.delta;
+          this.emit({ type: "text.updated", sessionId: acc.sessionId, partId, text: acc.text });
+          return;
+        }
+        const racc = this.reasoningStreams.get(partId);
+        if (racc) {
+          racc.text += d.delta;
+          this.emit({ type: "reasoning.updated", sessionId: racc.sessionId, partId, text: racc.text });
+        }
+        break;
+      }
+      case "session.idle": {
+        const sessionId = String(props.sessionID ?? "");
+        // The turn is over — its text/reasoning parts can no longer receive
+        // deltas, and its step count resets for the next turn.
+        for (const [partId, acc] of this.textStreams)
+          if (acc.sessionId === sessionId) this.textStreams.delete(partId);
+        for (const [partId, acc] of this.reasoningStreams)
+          if (acc.sessionId === sessionId) this.reasoningStreams.delete(partId);
+        this.stepParts.delete(sessionId);
+        this.emit({ type: "session.idle", sessionId });
+        break;
+      }
+      // Interactive requests — support V2 (this server) and the bare names.
+      case "question.v2.asked":
+      case "question.asked": {
+        const q = props as {
+          id?: string;
+          sessionID?: string;
+          questions?: Array<{
+            question: string;
+            header: string;
+            options?: Array<{ label: string; description?: string }>;
+            multiple?: boolean;
+            custom?: boolean;
+          }>;
+        };
+        this.emit({
+          type: "question.asked",
+          sessionId: String(q.sessionID ?? ""),
+          requestId: String(q.id ?? ""),
+          questions: (q.questions ?? []).map((it) => ({
+            question: it.question,
+            header: it.header,
+            options: it.options ?? [],
+            multiple: it.multiple,
+            custom: it.custom,
+          })),
+        });
+        break;
+      }
+      case "question.v2.replied":
+      case "question.v2.rejected":
+      case "question.replied":
+      case "question.rejected": {
+        const q = props as { requestID?: string; id?: string; sessionID?: string };
+        this.emit({
+          type: "question.resolved",
+          sessionId: String(q.sessionID ?? ""),
+          requestId: String(q.requestID ?? q.id ?? ""),
+        });
+        break;
+      }
+      case "permission.v2.asked":
+      case "permission.asked": {
+        // The V2 server names the fields `permission` + `patterns`;
+        // older payloads used `action` + `resources`. Accept both.
+        const p = props as {
+          id?: string;
+          sessionID?: string;
+          permission?: string;
+          patterns?: string[];
+          action?: string;
+          resources?: string[];
+        };
+        this.emit({
+          type: "permission.asked",
+          sessionId: String(p.sessionID ?? ""),
+          requestId: String(p.id ?? ""),
+          action: String(p.permission ?? p.action ?? "action"),
+          resources: p.patterns ?? p.resources ?? [],
+        });
+        break;
+      }
+      case "permission.v2.replied":
+      case "permission.replied": {
+        const p = props as { requestID?: string; id?: string; sessionID?: string };
+        this.emit({
+          type: "permission.resolved",
+          sessionId: String(p.sessionID ?? ""),
+          requestId: String(p.requestID ?? p.id ?? ""),
+        });
+        break;
+      }
+      case "session.error": {
+        // OpenCode nests the human-readable message at error.data.message.
+        // Keep the first line — OpenCode appends a stack trace to some errors.
+        this.emit({
+          type: "error",
+          sessionId: String(props.sessionID ?? ""),
+          message: errorText(props.error) ?? "session error",
+        });
+        break;
+      }
+      case "session.status": {
+        // A failed model call being retried server-side. OpenCode's retry
+        // policy has NO attempt cap (exponential backoff only), so these are
+        // the ONLY sign of life while a broken provider fails every attempt —
+        // dropping them leaves the UI on a bare "Working…" forever. ("busy"
+        // and "idle" statuses carry nothing session.idle doesn't already.)
+        const status = props.status as
+          | { type?: string; attempt?: number; message?: string; next?: number }
+          | undefined;
+        if (status?.type !== "retry") break;
+        this.emit({
+          type: "session.retry",
+          sessionId: String(props.sessionID ?? ""),
+          attempt: status.attempt ?? 0,
+          message: status.message ?? "provider error",
+          nextAt: status.next ?? 0,
+        });
+        break;
+      }
+      default:
+        break; // server.connected and others are ignored
+    }
+  }
 }
 
 // Type assertion: the engine structurally satisfies AgentRuntime. If the
 // interface changes and this engine isn't updated, tsc fails here.
-import type { AgentRuntime } from "../runtime";
 const _satisfiesAgentRuntime: AgentRuntime = null as unknown as OpenCodeHttpEngine;
 void _satisfiesAgentRuntime;
